# storage/api-docs/api-docs.json
{
    "openapi": "3.0.0",
    "info": {
        "title": "Vive Hidalgo API",
        "description": "API para la plataforma de turismo Vive Hidalgo",
        "contact": {
            "name": "Soporte Vive Hidalgo",
            "email": "admin@vivehidalgo.com"
        },
        "license": {
            "name": "MIT",
            "url": "https://opensource.org/licenses/MIT"
        },
        "version": "1.0.0"
    },
    "servers": [
        {
            "url": "http://localhost:8000",
            "description": "Main API Server"
        },
        {
            "url": "http://localhost:8000",
            "description": "API Server"
        }
    ],
    "paths": {
...
        "/api/v1/media/upload": {
            "post": {
                "tags": [
                    "Media"
                ],
                "summary": "Subir imagen",
                "description": "Sube una imagen y la asocia a un modelo específico",
                "operationId": "311248825de6d81ab0bcf48f911ede9a",
                "requestBody": {
                    "required": true,
                    "content": {
                        "multipart/form-data": {
                            "schema": {
                                "required": [
                                    "file",
                                    "imageable_type",
                                    "imageable_id"
                                ],
                                "properties": {
                                    "file": {
                                        "description": "Archivo de imagen",
                                        "type": "string",
                                        "format": "binary"
                                    },
                                    "imageable_type": {
                                        "description": "Tipo de modelo (App\\\\Models\\\\Destino, App\\\\Models\\\\Promocion, etc.)",
                                        "type": "string"
                                    },
                                    "imageable_id": {
                                        "description": "ID del modelo",
                                        "type": "integer"
                                    },
                                    "alt": {
                                        "description": "Texto alternativo de la imagen",
                                        "type": "string"
                                    },
                                    "is_main": {
                                        "description": "Indica si es la imagen principal",
                                        "type": "boolean"
                                    },
                                    "orden": {
                                        "description": "Orden de la imagen",
                                        "type": "integer"
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "201": {
                        "description": "Imagen subida exitosamente",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "success": {
                                            "type": "boolean",
                                            "example": true
                                        },
                                        "message": {
                                            "type": "string",
                                            "example": "Imagen subida exitosamente"
                                        },
                                        "data": {
                                            "$ref": "#/components/schemas/Imagen"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "400": {
                        "description": "Error de validación",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "success": {
                                            "type": "boolean",
                                            "example": false
                                        },
                                        "message": {
                                            "type": "string",
                                            "example": "Error de validación"
                                        },
                                        "errors": {
                                            "type": "object"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "403": {
                        "description": "No autorizado",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "success": {
                                            "type": "boolean",
                                            "example": false
                                        },
                                        "message": {
                                            "type": "string",
                                            "example": "No autorizado"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    }
                },
                "security": [
                    {
                        "sanctum": []
                    }
                ]
            }
        },
        "/api/v1/media/{imagen}": {
            "delete": {
                "tags": [
                    "Media"
                ],
                "summary": "Eliminar imagen",
                "description": "Elimina una imagen específica",
                "operationId": "d3a2d22039ac4e0e20beb1b161d5837c",
                "parameters": [
                    {
                        "name": "imagen",
                        "in": "path",
                        "description": "ID de la imagen",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Imagen eliminada exitosamente",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "success": {
                                            "type": "boolean",
                                            "example": true
                                        },
                                        "message": {
                                            "type": "string",
                                            "example": "Imagen eliminada exitosamente"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "404": {
                        "description": "Imagen no encontrada",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "success": {
                                            "type": "boolean",
                                            "example": false
                                        },
                                        "message": {
                                            "type": "string",
                                            "example": "Imagen no encontrada"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "403": {
                        "description": "No autorizado",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "success": {
                                            "type": "boolean",
                                            "example": false
                                        },
                                        "message": {
                                            "type": "string",
                                            "example": "No autorizado"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    }
                },
                "security": [
                    {
                        "sanctum": []
                    }
                ]
            }
        },
...
    },
    "components": {
        "schemas": {
...
            "Imagen": {
                "title": "Imagen",
                "required": [
                    "id",
                    "path",
                    "imageable_type",
                    "imageable_id"
                ],
                "properties": {
                    "id": {
                        "description": "ID único de la imagen",
                        "type": "integer",
                        "format": "int64"
                    },
                    "path": {
                        "description": "Ruta del archivo de imagen",
                        "type": "string"
                    },
                    "alt": {
                        "description": "Texto alternativo de la imagen",
                        "type": "string",
                        "nullable": true
                    },
                    "orden": {
                        "description": "Orden de la imagen",
                        "type": "integer"
                    },
                    "is_main": {
                        "description": "Indica si es la imagen principal",
                        "type": "boolean"
                    },
                    "url": {
                        "description": "URL completa de la imagen",
                        "type": "string"
                    },
                    "created_at": {
                        "type": "string",
                        "format": "date-time",
                        "readOnly": true
                    },
                    "updated_at": {
                        "type": "string",
                        "format": "date-time",
                        "readOnly": true
                    }
                },
                "type": "object"
            },
...
        },
        "securitySchemes": {
            "bearerAuth": {
                "type": "http",
                "scheme": "bearer"
            },
            "sanctum": {
                "type": "http",
                "bearerFormat": "JWT",
                "scheme": "bearer"
            }
        }
    },
...
}